Observables

Observables are heavily used in Output writers.

Various parts of Penning.jl (e.g. Output writers or callbacks) need to execute certain actions at specific time intervals. For that, a family of schedules were introduced that cover a variety of basic scheduling tasks:

PositionObservable

VelocityObservable

KineticEnergyObservable

Additionally, there are Higher-order observables, which are not associated to a specific physical quantity, but perform reducing or extraction operations on child observables. For example, you can use a SumObservable to observe the sum over all

VectorComponentObservable

The observe function

Penning.Observables.observe — Function

observe(obs, selection, setup)

Returns the value of an observable obs of a given selection.

The returned type depends on the specific observable.

PositionObservable

Penning.Observables.PositionObservable — Type

PositionObservable()

Defines an observable that returns the euclidian position vector of a particle.

VelocityObservable

Penning.Observables.VelocityObservable — Type

VelocityObservable()

Defines an observable that returns the euclidian velocity vector of a particle.

KineticEnergyObservable

Missing docstring.

Missing docstring for KineticEnergyObservable. Check Documenter's build log for details.

ElectrodeObservable

Missing docstring.

Missing docstring for ElectrodeObservable. Check Documenter's build log for details.

Higher-order observables

VectorComponentObservable

Penning.Observables.VectorComponentObservable — Type

VectorComponentObservable(child_observable, component)

Defines an observable that returns a single component (x, y, or z) of a child vector-observable specified by child_observable. The component is specified by the component argument. Valid components are the integers 1,2,3 or the symbols :x,:y,:z.

For example, if you only want to observe the x position of a particle, you can do so using the following code:

my_observable = VectorComponentObservable(PositionObservable(), :x)