Fields

In Penning.jl AbstractField

Trap Fields

IdealTrapField

Penning.Fields.IdealTrapField — Type

IdealTrapField(U₀, c₂, B₀)

Field of an ideal Penning trap, specified by potential depth U₀ in volts, quadratic field coefficient c₂ in 1/m² and magnetic field strength B₀ in Tesla. The magnetic field component is aligned in the $z$ direction.

\[\vec{E} = c_2\, U_0 \left( \begin{array}{c} x \\ y \\ -2z \end{array} \right) \qquad \vec{B} = \left( \begin{array}{c} 0 \\ 0 \\ B_0 \end{array} \right)\]

Excitation Fields

DipolarExcitationField

Penning.Fields.DipolarExcitationField — Type

DipolarExcitationField(ω, A)

Electric dipolar RF excitation field with angular frequency ω given in in rad/s. The amplitude and direction of the electric dipole field is specified by the three-dimensional vector argument A.

\[\vec{E} = \vec{A} \cdot \sin (\omega\, t) \qquad \vec{B} = \vec{0}\]

QuadrupolarExcitationField

Penning.Fields.QuadrupolarExcitationField — Type

QuadrupolarExcitationField(ω, A)

PlaneWaveExcitationField

Penning.Fields.PlaneWaveExcitationField — Type

PlaneWaveExcitationField(ω, E₀)

TECavityExcitationField

Missing docstring.

Missing docstring for TECavityExcitationField(::Integer, ::Integer, ::Integer, ::Number, ::Number). Check Documenter's build log for details.

TMCavityExcitationField

Missing docstring.

Missing docstring for TMCavityExcitationField(::Integer, ::Integer, ::Integer, ::Number, ::Number). Check Documenter's build log for details.