Particles

Particle collection

In Penning.jl, equally behaving particles are contained in a ParticleCollection

Penning.Particles.ParticleCollection — Type

Fields:

species: Instance of type Species, specifying the type (charge and mass) of particles in this collection

r: Position of the particles in this collection

v: Velocities of the particles in this collection

E: Used to store the electric field at the positions of the individual particles during simulation

E: Used to store the magnetic field at the positions of the individual particles during simulation

damping: 3-dimensional vector specifying an artificial damping constant for all particles in this collection

Species

Penning.Particles.Species — Type

Species

Parameter object for spcifying the mass m (in kg) and the charge q (in C) of a particle species.

Penning.Particles.Electron — Method

Electron()

Predefined particle species for the electron

Penning.Particles.Ion — Method

Ion(m, q)

Function for generating a particle species using units commonly used in atomic physics: Mass m is provided in atomic mass units and charge q is provided in units of (positive) elementary charge.

Example:

The following instantiates an ion with mass 187 u and charge state 28+

my_ion = Ion(187, 28)

Particle disttributions