Schedules

Various parts of Penning.jl (e.g. Output writers or callbacks) need to execute certain actions at specific time intervals. For that, a family of schedules were introduced that cover a variety of basic scheduling tasks:

TimeInterval

AlignedTimeInterval

IterationInterval

WallTimeInterval

SpecifiedTimes

Additionally, there are "composit schedules", that link different schedules together, to build complex time dependencies:

ConsecutiveIterations

AndSchedule

OrSchedule

TimeInterval

Penning.Utils.TimeInterval — Method

TimeInterval(interval)

Returns a callable TimeInterval that schedules periodic output or diagnostic evaluation on a interval of simulation time, as kept by model.clock. Pay attention that this schedule might not actuate in equidistant intervals if the passed interval is not properly aligned with the simulation time step. See AlignedTimeInterval for an aligned version.

AlignedTimeInterval

Penning.Utils.AlignedTimeInterval — Method

AlignedTimeInterval(interval, dt)

Similar to TimeInterval, but aligns the interval to multiples of dt. Actually, this function returns an IterationInterval in order to guarantee for exactly equidistant "actuation events".

IterationInterval

Penning.Utils.IterationInterval — Method

IterationInterval(interval; offset=0)

Returns a callable IterationInterval that "actuates" whenever the simulation iteration (modified by offset) is a multiple of interval. For example,

IterationInterval(100) actuates at iterations [100, 200, 300, ...].

IterationInterval(100, offset=-1) actuates at iterations [99, 199, 299, ...].

WallTimeInterval

Penning.Utils.WallTimeInterval — Method

WallTimeInterval(interval; start_time = time_ns() * 1e-9)

Returns a callable WallTimeInterval that actuates on a interval of "wall time" while a simulation runs, in units of seconds. The "wall time" is the actual real world time in seconds, as kept by an actual or hypothetical clock hanging on your wall. The keyword argument start_time can be used to specify a starting wall time other than the moment WallTimeInterval is constructed.

SpecifiedTimes

Penning.Utils.SpecifiedTimes — Method

SpecifiedTimes(times)

Returns a callable SpecifiedTimes that actuates at predefined simulation times. Pass times either by individual arguments SpecifiedTimes(0.0, 3.14, 8.9) or as a list SpecifiedTimes([0.0, 3.14, 8.9])

ConsecutiveIterations

Missing docstring.

Missing docstring for ConsecutiveIterations(parent_schedule). Check Documenter's build log for details.

AndSchedule

Penning.Utils.AndSchedule — Method

AndSchedule(child_schedule_1, child_schedule_2, other_child_schedules...)

Return a schedule that actuates when all child_schedules actuate.

OrSchedule

Penning.Utils.OrSchedule — Method

OrSchedule(child_schedule_1, child_schedule_2, other_child_schedules...)

Return a schedule that actuates when any of the child_schedules actuates.